Strip of the query part when building the verification URI
The query part of the request URI was copied to the verification URI - which is - in my opinion - not the expected behaviour.

According to https://datatracker.ietf.org/doc/html/rfc8628#section-3.1

"Parameters sent without a value MUST be treated as if they were omitted from the request.  The authorization server MUST ignore unrecognized request parameters.  Request and response parameters MUST NOT be included more than once."

Passing through the request parameters is - IMHO - a different behaviour than ignoring them.